Incorporated suggestions
